fix: stuck at leaving main
Before jump to main, no return address are stored to stack. So when leaving
from main, it attemps pop return address from stack, stack pointer is
the init value at that time, it is the end of ram, the same as uart
address, so the pop instruction functions as read a char from uart, it
will stuck there. Change the jump to entry function to bl instructions
to fix this problem

After return from main, add code to output "end" to tell that the
program is teminated, then a deadloop.