feat: implement cmp instructions to support if else
Implement LT,LE,GT,GE instructions to support if and else

Modify implementation of cmp instructions. Because cmp instructons in c4
will save the cmp result to a, and BZ and BNZ make decisions base on
value in a. But in qemu, cmp result is stored in spr.z, a register
invisble for c language. So add a instruction to save spr.z to a before
the jump action in BZ and BNZ instructions. This will make if (a > 0)
and if (a) both work well